requirements.txt
flask==3.1.3
psycopg2-binary==2.9.11
python-dotenv==1.2.2
werkzeug==3.1.6
gunicorn==23.0.0
flask-mail==0.10.0
openpyxl==3.1.5
reportlab==4.4.10
